Change Controllers to Support Version
For each of the controllers, add a version field to each action which
renders a response - then, integrate the version aspect directly into
the application. That way, if we ever have more versions, we can just
access a different version of the page, and it will reflect on the
rating inside the database.